+++ src/mr_recon/recons.py
@@ -3,6 +3,7 @@ import time
 
 from tqdm import tqdm
 from typing import Optional
+from mr_recon.dtypes import complex_dtype
 from mr_recon.linops import linop
 from mr_recon.utils import np_to_torch, torch_to_np
 from mr_recon.algs import (
@@ -49,7 +50,7 @@ def min_norm_recon(A: linop,
 
     # Estimate largest eigenvalue so that lambda max of AHA is 1
     if max_eigen is None:
-        x0 = torch.randn(A.ishape, dtype=torch.complex64, device=device)
+        x0 = torch.randn(A.ishape, dtype=complex_dtype, device=device)
         _, max_eigen = power_method_operator(A.normal, x0, verbose=verbose)
         max_eigen *= 1.01
 
@@ -58,7 +59,7 @@ def min_norm_recon(A: linop,
 
     # Run CG
     y = conjugate_gradient(AHA=AAH, 
-                           AHb=ksp.type(torch.complex64),
+                           AHb=ksp.type(complex_dtype),
                            lamda_l2=lamda_l2,
                            num_iters=max_iter,
                            verbose=verbose)
@@ -108,13 +109,13 @@ def CG_SENSE_recon(A: linop,
 
     # Estimate largest eigenvalue so that lambda max of AHA is 1
     if max_eigen is None:
-        x0 = torch.randn(A.ishape, dtype=torch.complex64, device=device)
+        x0 = torch.randn(A.ishape, dtype=complex_dtype, device=device)
         _, max_eigen = power_method_operator(A.normal, x0, verbose=verbose)
         max_eigen *= 1.01
     
     # Starting with AHb
     start = time.perf_counter()
-    y = ksp.type(torch.complex64)
+    y = ksp.type(complex_dtype)
     AHb = A.adjoint(y) / (max_eigen ** 0.5)
     end = time.perf_counter()
     if verbose:
@@ -222,13 +223,13 @@ def FISTA_recon(A: linop,
 
     # Estimate largest eigenvalue so that lambda max of AHA is 1
     if max_eigen is None:
-        x0 = torch.randn(A.ishape, dtype=torch.complex64, device=device)
+        x0 = torch.randn(A.ishape, dtype=complex_dtype, device=device)
         _, max_eigen = power_method_operator(A.normal, x0, verbose=verbose)
         max_eigen *= 1.01
     
     # Starting with AHb
     start = time.perf_counter()
-    y = ksp.type(torch.complex64)
+    y = ksp.type(complex_dtype)
     AHb = A.adjoint(y) / (max_eigen ** 0.5)
     end = time.perf_counter()
     if verbose:
